requirements.txt
asgiref==3.5.2
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
Deprecated==1.2.13
Django==4.1.3
django-auth-ldap3-ad==1.6.33
django-countries==7.4.2
django-storages==1.13.1
djangorestframework==3.14.0
Flask==2.2.2
gitdb==4.0.9
github3.py==3.2.0
GitPython==3.1.29
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
ldap3==2.9.1
lxml==4.9.1
MarkupSafe==2.1.1
mysqlclient==2.1.1
paginate==0.5.6
paginate-sqlalchemy==0.3.1
paramiko==2.12.0
Pillow==9.3.0
pyadaptivecards==0.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyGithub==1.57
PyJWT==2.6.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
python-ldap==3.4.3
python-pptx==0.6.21
pytz==2022.6
requests==2.28.1
requests-toolbelt==0.10.1
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.42
sqlparse==0.4.3
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
XlsxWriter==3.0.3
